+++ src/module/applications/hud/_module.mjs
@@ -1,1 +1,0 @@
-export { default as DrawSteelTokenHUD } from "./token-hud.mjs";