+++ src/module/applications/sheets/actor-sheet.mjs
@@ -175,7 +175,7 @@ export default class DrawSteelActorSheet extends api.HandlebarsApplicationMixin(
         break;
       case "biography":
         context.languages = this._getLanguages();
-        context.enrichedBiography = await TextEditor.enrichHTML(
+        context.enrichedBiography = await CONFIG.ux.TextEditor.enrichHTML(
           this.actor.system.biography.value,
           {
             secrets: this.document.isOwner,
@@ -183,7 +183,7 @@ export default class DrawSteelActorSheet extends api.HandlebarsApplicationMixin(
             relativeTo: this.actor,
           },
         );
-        context.enrichedGMNotes = await TextEditor.enrichHTML(
+        context.enrichedGMNotes = await CONFIG.ux.TextEditor.enrichHTML(
           this.actor.system.biography.gm,
           {
             secrets: this.document.isOwner,
@@ -228,7 +228,7 @@ export default class DrawSteelActorSheet extends api.HandlebarsApplicationMixin(
     const actorMovement = this.actor.system.movement;
     const flying = actorMovement.types.has("fly");
     const movementList = Array.from(actorMovement.types).map(m => {
-      let label = game.i18n.localize(ds.CONFIG.movementTypes[m]?.label ?? m);
+      let label = game.i18n.localize(CONFIG.Token.movement.actions[m]?.label ?? m);
       if ((m === "teleport") && (actorMovement.teleport !== actorMovement.value)) label += " " + actorMovement.teleport;
       return label;
     });
@@ -236,7 +236,9 @@ export default class DrawSteelActorSheet extends api.HandlebarsApplicationMixin(
     return {
       flying,
       list: formatter.format(movementList),
-      options: Object.entries(ds.CONFIG.movementTypes).filter(([_key, a]) => a.speedOption).map(([value, { label }]) => ({ value, label })),
+      options: Object.entries(CONFIG.Token.movement.actions)
+        .filter(([key, _action]) => ds.CONFIG.speedOptions.includes(key))
+        .map(([value, { label }]) => ({ value, label })),
     };
   }